# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RelaySMS Publisher",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/platforms": {
      "get": {
        "tags": [
          "Platforms"
        ],
        "summary": "List platforms",
        "description": "Platforms users can link: those with an installed, enabled adapter.",
        "operationId": "get_platforms_v1_platforms_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 50,
                  "pattern": "^[a-zA-Z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by platform name",
              "title": "Name"
            },
            "description": "Filter by platform name"
          },
          {
            "name": "proto_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by protocol ID",
              "title": "Proto Id"
            },
            "description": "Filter by protocol ID"
          },
          {
            "name": "cat_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category ID",
              "title": "Cat Id"
            },
            "description": "Filter by category ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlatformManifest"
                  },
                  "title": "Response Get Platforms V1 Platforms Get"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/platforms/adapters": {
      "get": {
        "tags": [
          "Platform Adapters"
        ],
        "summary": "List adapters",
        "description": "Every installed adapter, disabled ones included. Scope: platforms:read.",
        "operationId": "list_adapters_v1_platforms_adapters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlatformAdapterInfo"
                  },
                  "type": "array",
                  "title": "Response List Adapters V1 Platforms Adapters Get"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks platforms:read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": [
              "platforms:read"
            ]
          },
          {
            "SessionCookie": [
              "platforms:read"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "Platform Adapters"
        ],
        "summary": "Install adapter",
        "description": "Queues an install; poll the job at Location. Administrators only.",
        "operationId": "install_adapter_v1_platforms_adapters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdapterInstall"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdapterJobInfo"
                }
              }
            }
          },
          "400": {
            "description": "source_url isn't a GitHub repository of an org in PLATFORMS_GITHUB_ORGS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks platforms:write, or a cookie request came from an origin that isn't allowed. Also when the credential isn't an administrator.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Already installed, or an install of it is already running.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The job queue is unavailable. Try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": [
              "platforms:write"
            ]
          },
          {
            "SessionCookie": [
              "platforms:write"
            ]
          }
        ]
      }
    },
    "/v1/platforms/adapters/{adapter_id}": {
      "get": {
        "tags": [
          "Platform Adapters"
        ],
        "summary": "Get adapter",
        "description": "Returns the ETag to send as If-Match. Scope: platforms:read.",
        "operationId": "get_adapter_v1_platforms_adapters__adapter_id__get",
        "security": [
          {
            "HTTPBasic": [
              "platforms:read"
            ]
          },
          {
            "SessionCookie": [
              "platforms:read"
            ]
          }
        ],
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 36,
              "pattern": "^[A-Za-z0-9._-]+$",
              "title": "Adapter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformAdapterInfo"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks platforms:read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No adapter with that ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Platform Adapters"
        ],
        "summary": "Enable or disable",
        "description": "Needs If-Match. Scope: platforms:write.\n\nA disabled adapter is hidden from users but still revokes tokens.",
        "operationId": "update_adapter_v1_platforms_adapters__adapter_id__patch",
        "security": [
          {
            "HTTPBasic": [
              "platforms:write"
            ]
          },
          {
            "SessionCookie": [
              "platforms:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 36,
              "pattern": "^[A-Za-z0-9._-]+$",
              "title": "Adapter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformAdapterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformAdapterInfo"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks platforms:write, or a cookie request came from an origin that isn't allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No adapter with that ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Changed by another request at the same moment. Reload and try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "If-Match isn't the current ETag: it changed since you loaded it. Reload and try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "428": {
            "description": "No If-Match header. Send the ETag from the last GET.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Platform Adapters"
        ],
        "summary": "Uninstall adapter",
        "description": "Needs If-Match. Scope: platforms:write.\n\nRefused while accounts are linked through it; disable it instead.",
        "operationId": "delete_adapter_v1_platforms_adapters__adapter_id__delete",
        "security": [
          {
            "HTTPBasic": [
              "platforms:write"
            ]
          },
          {
            "SessionCookie": [
              "platforms:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 36,
              "pattern": "^[A-Za-z0-9._-]+$",
              "title": "Adapter Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks platforms:write, or a cookie request came from an origin that isn't allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No adapter with that ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Accounts are linked through it, or it changed at the same moment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "If-Match isn't the current ETag: it changed since you loaded it. Reload and try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "428": {
            "description": "No If-Match header. Send the ETag from the last GET.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/platforms/adapters/{adapter_id}/update": {
      "post": {
        "tags": [
          "Platform Adapters"
        ],
        "summary": "Update adapter",
        "description": "Queues a move to a version tag; poll the job at Location. Administrators only.",
        "operationId": "update_adapter_version_v1_platforms_adapters__adapter_id__update_post",
        "security": [
          {
            "HTTPBasic": [
              "platforms:write"
            ]
          },
          {
            "SessionCookie": [
              "platforms:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 36,
              "pattern": "^[A-Za-z0-9._-]+$",
              "title": "Adapter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdapterUpgrade"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdapterJobInfo"
                }
              }
            }
          },
          "400": {
            "description": "The adapter's source isn't a GitHub repository of an org in PLATFORMS_GITHUB_ORGS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks platforms:write, or a cookie request came from an origin that isn't allowed. Also when the credential isn't an administrator.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No adapter with that ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "An install or update of it is already running.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The job queue is unavailable. Try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdapterInstall": {
        "properties": {
          "source_url": {
            "type": "string",
            "maxLength": 255,
            "title": "Source Url",
            "description": "https://github.com/<org>/<repo>, an org in PLATFORMS_GITHUB_ORGS."
          },
          "tag": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag",
            "description": "A version tag such as v1.2.0. Defaults to the newest one."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_url"
        ],
        "title": "AdapterInstall"
      },
...
      "AdapterUpgrade": {
        "properties": {
          "tag": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag",
            "description": "A version tag such as v1.2.0. Defaults to the newest one."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AdapterUpgrade"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "What went wrong, safe to show to users."
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The body of every error response."
      },
...
      "PlatformAdapterUpdate": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "PlatformAdapterUpdate"
      },
      "PlatformManifest": {
        "properties": {
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "proto_id": {
            "type": "integer",
            "title": "Proto Id",
            "description": "0 = oauth2, 1 = pnba"
          },
          "cat_id": {
            "type": "integer",
            "title": "Cat Id",
            "description": "0 = email, 1 = message, 2 = text, 3 = bridge"
          },
          "auth_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Provider"
          },
          "supports_offline_first": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supports Offline First"
          },
          "icon_svg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Svg"
          },
          "icon_png": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Png"
          }
        },
        "type": "object",
        "required": [
          "display_name",
          "name",
          "proto_id",
          "cat_id"
        ],
        "title": "PlatformManifest"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "SessionCookie": {
        "type": "apiKey",
        "description": "Set by POST /v1/auth/login.",
        "in": "cookie",
        "name": "relaysms_session"
      }
    }
  }
}